requirements.txt
asn1crypto==0.24.0
begins==0.9
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.3.1
elasticsearch==5.5.3
Flask==1.0.2
Flask-Cors==3.0.6
gevent==1.3.6
greenlet==0.4.15
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pycparser==2.19
pymongo==3.7.1
PyMySQL==0.9.2
requests==2.19.1
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
